Adjust whitespace to better match linux style.
The idea is to make haxv6 as close to linux code style as possible since I want
to use haxv6 as a playground to develop myself as a linux hacker. Also, the
linux style guide provides consistency with some pretty sane defaults and allows
me to use the checkpatch.pl script to find issues.