Remove unnecessary braces.
In many code conventions braces are always used even when unnecessary to avoid
mistakes like:

if (foo)
	bar();
else
	baz();
	blargh();

However this convention is not used in linux. Most of the time this kind of
mistake is visually obvious and the braces add noise so thre is an argument for
it.

Via linux checkpatch.pl.