Move scripts to their own directory.
This is part of a larger effort to add some directory structure and thus make
working with haxv6 a little easier.

This, in turn, is part of an even larger effort to align haxv6 with linux as
much as possible so it's more useful for my overall purpose of having a
as-best-I-can-linux-lite to fiddle with.

But we need to crawl before we can walk, so let's get the trivial stuff out of
the way first :)