Move code to linux-like folder structure.
I much prefer this to having all code in 1 single directory. It also lays
foundations for further converting the codebase to better resemble linux in less
trivial ways.

The generated files are still all placed in the root directory, so things remain
messy that way. I plan to fix that and therefore generate less noise, as well as
having separate Makefile's for separate components.